*: extract common merge functionality
